requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiomultiprocess==0.9.1
aiosignal==1.3.1
annotated-types==0.7.0
argcomplete==3.5.1
async-timeout==5.0.1
attrs==24.2.0
Automat==20.2.0
Babel==2.8.0
bc-detect-secrets==1.5.22
bc-jsonpath-ng==1.6.1
bc-python-hcl2==0.4.2
bcrypt==3.2.0
beartype==0.19.0
beautifulsoup4==4.12.3
blinker==1.4
boolean.py==4.0
boto3==1.35.49
botocore==1.35.63
cached-property==2.0.1
cachetools==5.5.0
certifi==2020.6.20
cffi==1.17.1
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.4.0
checkov==3.2.301
click==8.1.7
click-option-group==0.5.6
cloud-init==24.3.1
cloudsplaining==0.7.0
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.7
configobj==5.0.6
constantly==15.1.0
contextlib2==21.6.0
cryptography==3.4.8
cyclonedx-python-lib==6.4.4
dbus-python==1.2.18
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==7.1.0
dockerfile-parse==2.0.1
dpath==2.1.3
filelock==3.16.1
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
httplib2==0.20.2
hyperlink==21.0.0
identify==2.6.2
idna==3.3
importlib_metadata==7.2.1
incremental==21.3.0
isodate==0.7.2
jeepney==0.7.1
Jinja2==3.0.3
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
junit-xml==1.9
keyring==23.5.0
lark==1.2.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
license-expression==30.4.0
Markdown==3.7
MarkupSafe==2.0.1
more-itertools==8.10.0
multidict==6.1.0
netifaces==0.11.0
networkx==2.6.3
nodeenv==1.9.1
numpy==2.1.3
oauthlib==3.2.0
openai==0.28.1
orjson==3.10.11
packageurl-python==0.13.4
packaging==23.2
platformdirs==4.3.6
ply==3.11
policy_sentry==0.13.1
pre_commit==4.0.1
prettytable==3.12.0
propcache==0.2.0
psycopg2-binary==2.9.10
py-serializable==1.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycares==4.4.0
pycep-parser==0.5.1
pycparser==2.22
pycurl==7.44.1
pydantic==2.9.2
pydantic_core==2.23.4
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pyston==2.3.5
pyston-autoload==2.3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
pytz==2022.1
PyYAML==6.0.2
rdflib==7.1.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.21.0
rustworkx==0.13.2
s3transfer==0.10.3
schema==0.7.5
SecretStorage==3.3.1
semantic-version==2.10.0
service-identity==18.1.0
six==1.16.0
smmap==5.0.1
sortedcontainers==2.4.0
soupsieve==2.6
spdx-tools==0.8.3
systemd-python==234
tabulate==0.9.0
termcolor==2.3.0
tqdm==4.67.0
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unidiff==0.7.5
uritools==4.0.3
urllib3==1.26.5
virtualenv==20.27.1
wadllib==1.3.6
wcwidth==0.2.13
xmltodict==0.14.2
yarl==1.17.2
zipp==1.0.0
zope.interface==5.4.0
